flaskapp/requirements.txt
Flask>=2.2.2
networkx
requests
requests-oauthlib
google-cloud-bigquery
pandas
